Refactor project to components
