+++ src/App.js
@@ -1,4 +1,4 @@
-import TopBar from "./topbar/TopBar";
+import TopBar from "./components/topbar/TopBar";
 import Home from "./pages/home/Home";
