+++ src/pages/home/Home.jsx
@@ -1,6 +1,6 @@
-import Header from "../../header/Header";
-import Posts from "../../posts/Posts";
-import SideBar from "../../sidebar/SideBar";
+import Header from "../../components/header/Header";
+import Posts from "../../components/posts/Posts";
+import SideBar from "../../components/sidebar/SideBar";
 import "./home.css";
